lecture ok

+++ define.h
@@ -17,10 +17,10 @@
 
 typedef struct {
 
-    float ACR;
-    float DCR;
-    float ACIR;
-    float DCIR;
+    int ACR;
+    int DCR;
+    int ACIR;
+    int DCIR;
 } absorp;
 
 //structure contenant les valeurs du pouls et de SPO2
